+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Use `button` for close button in `Doggo.modal/1`.
 - Add `dismissable` attribute to `Doggo.modal/1`.
 - Remove `role` from `button_link/1`, add `class`.
+- Rename `Doggo.drawer/1` slots to `header`, `main`, and `footer`.
 - Depend on `phoenix_storybook ~> 0.6.0`.
 
 ## [0.4.0] - 2023-12-31